Add local build option that requires restate dependencies
This commit introduces the E2E_LOCAL_BUILD env variable that activates
the local build mode. In the local build mode, Gradle expects all restate
dependencies (java-sdk) to be locally available under ../java-sdk. It will
then include these dependencies in the build as a composite build. This
ensures that local changes to this repository can be easily tested locally.

This fixes #18.