Cleanup snapshot artifacts by deleting snapshot version before publishing
This is a work around for the problem that Github does not allow to delete
artifacts of a package version. Only the whole package version can be deleted.
Since snapshot artifacts are published under the same snapshot version, this
commit makes sure that we only keep the latest one.

This fixes #39.